astu_yearbook/requirements.txt
Django==5.2.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.4.0
python-multipart==0.0.6
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
django-ratelimit==4.1.0
bleach==6.1.0
django-csp==3.8

# Database drivers
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching
django-redis==5.4.0
redis==5.0.1

# Monitoring and logging
sentry-sdk==1.40.0



